+++ backend/src/main/java/com/protosirius/backend/repository/MeasureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.protosirius.backend.entity.Measure;
 
-public interface MeasureRepository extends JpaRepository<Measure, Integer> {
+public interface MeasureRepository extends JpaRepository<Measure, Long> {
 
 }
